+++ src/protocol_level_tokens/mod.rs
@@ -7,11 +7,13 @@ mod token_account_info;
 pub mod token_client;
 mod token_event;
 mod token_info;
+mod token_reject_reason;
 
 pub use protocol_level_tokens::*;
 pub use token_account_info::*;
 pub use token_event::*;
 pub use token_info::*;
+pub use token_reject_reason::*;
 
 // gRPC type conversions for the types which are define as part of
 // the `concordium-base` crate.
@@ -99,23 +101,20 @@ impl TryFrom<generated::plt::token_event::Event> for TokenEventDetails {
     }
 }
 
-impl TryFrom<generated::plt::TokenModuleEvent> for TokenModuleEventEnum {
+impl TryFrom<generated::plt::TokenModuleEvent> for EncodedTokenModuleEvent {
     type Error = tonic::Status;
 
     fn try_from(event: generated::plt::TokenModuleEvent) -> Result<Self, Self::Error> {
-        let event_type = TokenModuleEventType::try_from_type_discriminator(
-            &event
+        Ok(Self {
+            event_type: event
                 .r#type
                 .try_into()
-                .map_err(|err: TypeFromStringError| tonic::Status::internal(err.to_string()))?,
-        )
-        .map_err(|err| tonic::Status::internal(err.to_string()))?;
-
-        Self::decode_event(
-            event_type,
-            &event.details.expect("Event details must exist").into(),
-        )
-        .map_err(|err| tonic::Status::internal(err.to_string()))
+                .map_err(|e: TypeFromStringError| tonic::Status::invalid_argument(e.to_string()))?,
+            details: event
+                .details
+                .ok_or(tonic::Status::invalid_argument("Event details must exist"))?
+                .into(),
+        })
     }
 }
 
@@ -159,25 +158,36 @@ impl TryFrom<generated::plt::TokenHolder> for TokenHolder {
     }
 }
 
-impl TryFrom<generated::plt::TokenModuleRejectReason> for TokenModuleRejectReasonEnum {
+impl TryFrom<generated::plt::TokenModuleRejectReason> for EncodedTokenModuleRejectReason {
     type Error = tonic::Status;
 
     fn try_from(value: generated::plt::TokenModuleRejectReason) -> Result<Self, Self::Error> {
-        let reject_reason_type = TokenModuleRejectReasonType::try_from_type_discriminator(
-            &value
+        Ok(Self {
+            token_id: value
+                .token_id
+                .ok_or(tonic::Status::invalid_argument("missing token id"))?
+                .try_into()?,
+            reason_type: value
                 .r#type
                 .try_into()
-                .map_err(|err: TypeFromStringError| tonic::Status::internal(err.to_string()))?,
-        )
-        .map_err(|err| tonic::Status::internal(err.to_string()))?;
+                .map_err(|e: TypeFromStringError| tonic::Status::invalid_argument(e.to_string()))?,
+            details: value.details.map(|c| c.into()),
+        })
+        // let reject_reason_type = TokenModuleRejectReasonType::try_from_type_discriminator(
+        //     &value
+        //         .r#type
+        //         .try_into()
+        //         .map_err(|err: TypeFromStringError| tonic::Status::internal(err.to_string()))?,
+        // )
+        // .map_err(|err| tonic::Status::internal(err.to_string()))?;
 
-        Self::decode_reject_reason(
-            reject_reason_type,
-            &value
-                .details
-                .expect("Reject reason details must be knwon")
-                .into(),
-        )
-        .map_err(|err| tonic::Status::internal(err.to_string()))
+        // Self::decode_reject_reason(
+        //     reject_reason_type,
+        //     &value
+        //         .details
+        //         .expect("Reject reason details must be knwon")
+        //         .into(),
+        // )
+        // .map_err(|err| tonic::Status::internal(err.to_string()))
     }
 }